+++ src/app/db/WDCReader.ts
@@ -336,15 +336,6 @@ export default class WDCReader {
 
 		// Ensure we've read the expected amount of pallet data.
 		assert.strictEqual(data.offset, prevPos + palletDataSize, 'Read incorrect amount of pallet data');
-
-		// New WDC4 chunk: TODO read
-		if (wdcVersion === 4) {
-			for (let sectionIndex = 0; sectionIndex < sectionCount - 1; sectionIndex++) {
-				const entryCount = data.readUInt32();
-				data.move(entryCount * 4);
-			}
-		}
-
 		prevPos = data.offset;
 
 		// char common_data[header.common_data_size];
@@ -361,6 +352,14 @@ export default class WDCReader {
 
 		// Ensure we've read the expected amount of common data.
 		assert.strictEqual(data.offset, prevPos + commonDataSize, 'Read incorrect amount of common data');
+
+		// New WDC4 chunk: TODO read
+		if (wdcVersion === 4) {
+			for (let sectionIndex = 0; sectionIndex < sectionCount - 1; sectionIndex++) {
+				const entryCount = data.readUInt32();
+				data.move(entryCount * 4);
+			}
+		}
 
 		// data_sections[header.section_count];
 		const sections = new Array(sectionCount);